backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
sqlalchemy==2.0.36
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
openai==1.54.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
